Feat: added a warning beforre deleting an entry

+++ src/components/Buttons/Delete/deleteCustomer.tsx
@@ -19,7 +19,9 @@ export const DeleteCustomerButton = ({
   );
 
   const handleDelete = () => {
-    dispatch(deleteCustomerThunk(id));
+    if (window.confirm("Are you sure you wanna delete this record?")) {
+      dispatch(deleteCustomerThunk(id));
+    }
   };
   return (
     <Button
